+++ components/AllCoinsGallery.tsx
@@ -11,12 +11,13 @@ type TProps = {
 export const AllCoinsGallery = ({ className }: TProps) => {
   return (
     <div className="">
-      <h1 className="font-title text-6xl flex justify-center">
+      <h1 className="font-title text-3xl flex justify-center">
         Last Tokens
       </h1>
       <div
         className={cn(
-          "font-body flex flex-wrap items-stretch max-md:place-content-center py-10 justify-center",
+          // "font-body flex flex-wrap items-stretch py-10 justify-center w-full",
+          "font-body w-full flex flex-wrap",
           className
         )}
       >
